Side quest: Disambiguate reference to code listing.
The phrase 'what you defined in posts.sdl.js:' ends in a colon,
followed by a code listing of services/posts/posts.js. It leads
the reader to believe that the source that follows is of posts.js,
rather than the source of posts.sdl.js.

Remove the colon to end the sentence, disambiguating the text
from the code listing that follows.