[Tutorial] Updates diff on in Saving Data section
Adds additional line highlighting on the "Display Server Errors" section
on the "Saving Data" page of the tutorial.

The section above mentiones updating the `Form` component:

> We'll also pass error to <Form> so it can setup a context

but the line with the `Form` component was not highlighted. This commit
highlights the line with the `Form` component to make it more obvious
this line needs to be updated as well.